Inline RocStructType back in
